\name{medseqr-package}
\alias{medseqr-package}
\alias{medseqr}
\docType{package}
\title{
  Utilities for analysis of biomedical data
}
\description{
  This package contains tools and helper functions to perform (usually small)
  manipulations of data structures such as data frames, matrices and lists,
  assist in gene expression data analysis and provide specific functionality
  to analysis scripts.
}
\details{
  \tabular{ll}{
    Package: \tab medseqr\cr
    Type: \tab Package\cr
    Version: \tab 0.1-0\cr
    License: \tab MIT License\cr
    LazyLoad: \tab yes\cr
  }
  The package contains small helper functions and utilities used commonly in data analysis workflows.
}
\keyword{ package }
